# colflow: more minor optimizations and pooling on the read path