--- a/pkg/sql/colcontainer/diskqueue.go
+++ b/pkg/sql/colcontainer/diskqueue.go
@@ -274,9 +274,10 @@ const (
 type DiskQueueCfg struct {
 	// FS is the filesystem interface to use.
 	FS fs.FS
-	// Path is where the temporary directory that will contain this DiskQueue's
-	// files should be created. The directory name will be a UUID.
-	Path string
+	// GetPath returns where the temporary directory that will contain this
+	// DiskQueue's files has been created. The directory name will be a UUID.
+	// Note that the directory is created lazily on the first call to GetPath.
+	GetPath func(context.Context) string
 	// CacheMode defines the way a DiskQueue should use its cache. Refer to the
 	// comment of DiskQueueCacheModes for more information.
 	CacheMode DiskQueueCacheMode
@@ -286,10 +287,6 @@ type DiskQueueCfg struct {
 	// MaxFileSizeBytes is the maximum size an on-disk file should reach before
 	// rolling over to a new one.
 	MaxFileSizeBytes int
-
-	// OnNewDiskQueueCb is an optional callback function that will be called when
-	// NewDiskQueue is called.
-	OnNewDiskQueueCb func()
 
 	// TestingKnobs are used to test the queue implementation.
 	TestingKnobs struct {
@@ -350,9 +347,6 @@ func newDiskQueue(
 	if err := cfg.EnsureDefaults(); err != nil {
 		return nil, err
 	}
-	if cfg.OnNewDiskQueueCb != nil {
-		cfg.OnNewDiskQueueCb()
-	}
 	d := &diskQueue{
 		dirName:          uuid.FastMakeV4().String(),
 		typs:             typs,
@@ -366,7 +360,7 @@ func newDiskQueue(
 	if d.cfg.CacheMode != DiskQueueCacheModeDefault {
 		d.writeBufferLimit = d.cfg.BufferSizeBytes / 2
 	}
-	if err := cfg.FS.MkdirAll(filepath.Join(cfg.Path, d.dirName)); err != nil {
+	if err := cfg.FS.MkdirAll(filepath.Join(cfg.GetPath(ctx), d.dirName)); err != nil {
 		return nil, err
 	}
 	// rotateFile will create a new file to write to.
@@ -413,7 +407,7 @@ func (d *diskQueue) Close(ctx context.Context) error {
 	if err := d.CloseRead(); err != nil {
 		return err
 	}
-	if err := d.cfg.FS.RemoveAll(filepath.Join(d.cfg.Path, d.dirName)); err != nil {
+	if err := d.cfg.FS.RemoveAll(filepath.Join(d.cfg.GetPath(ctx), d.dirName)); err != nil {
 		return err
 	}
 	totalSize := int64(0)
@@ -438,7 +432,7 @@ func (d *diskQueue) Close(ctx context.Context) error {
 // any file (i.e. during initialization). This will simply create the first file
 // to write to.
 func (d *diskQueue) rotateFile(ctx context.Context) error {
-	fName := filepath.Join(d.cfg.Path, d.dirName, strconv.Itoa(d.seqNo))
+	fName := filepath.Join(d.cfg.GetPath(ctx), d.dirName, strconv.Itoa(d.seqNo))
 	f, err := d.cfg.FS.CreateWithSync(fName, bytesPerSync)
 	if err != nil {
 		return err