**colflow: evaluate temp storage path lazily**

This commit makes the evaluation of the temp storage path in the
vectorized flow lazy which allows us to not create a uuid in case of
a local flow when the spilling to disk doesn't occur. This shows about
1% improvement on the micro benchmark of the flow setup. Additionally,
the directory creation is now done when its path is requested for the
first time which allows us to remove one callback from the config.

Release note: None

**colflow: more pooling of objects during the flow setup**

This commit keeps the references to several slices and a map that are
used during the flow setup which allows us to remove several
allocations.

Release note: None