@yuzefovich 2020-11-09
Microbenchmarks:
- first commit
```
name                                          old time/op    new time/op    delta
FlowSetup/vectorize=true/distribute=true-24     19.3µs ± 4%    19.1µs ± 5%    ~     (p=0.174 n=19+19)
FlowSetup/vectorize=true/distribute=false-24    16.9µs ± 3%    16.8µs ± 2%  -0.90%  (p=0.044 n=20+20)

name                                          old alloc/op   new alloc/op   delta
FlowSetup/vectorize=true/distribute=true-24     33.3kB ± 0%    33.2kB ± 0%  -0.35%  (p=0.000 n=19+18)
FlowSetup/vectorize=true/distribute=false-24    29.9kB ± 0%    29.7kB ± 0%  -0.37%  (p=0.000 n=16+16)

name                                          old allocs/op  new allocs/op  delta
FlowSetup/vectorize=true/distribute=true-24        324 ± 0%       321 ± 0%  -0.93%  (p=0.000 n=19+20)
FlowSetup/vectorize=true/distribute=false-24       293 ± 0%       290 ± 0%  -1.02%  (p=0.000 n=17+17)
```
- second commit
```
name                                          old time/op    new time/op    delta
FlowSetup/vectorize=true/distribute=true-24     19.4µs ± 6%    19.0µs ± 4%    ~     (p=0.051 n=20+19)
FlowSetup/vectorize=true/distribute=false-24    16.8µs ± 3%    16.8µs ± 4%    ~     (p=0.851 n=19+17)

name                                          old alloc/op   new alloc/op   delta
FlowSetup/vectorize=true/distribute=true-24     33.2kB ± 0%    33.2kB ± 0%  -0.27%  (p=0.000 n=18+16)
FlowSetup/vectorize=true/distribute=false-24    29.8kB ± 0%    29.7kB ± 1%  -0.20%  (p=0.003 n=18+19)

name                                          old allocs/op  new allocs/op  delta
FlowSetup/vectorize=true/distribute=true-24        322 ± 0%       318 ± 0%  -1.24%  (p=0.000 n=19+20)
FlowSetup/vectorize=true/distribute=false-24       291 ± 0%       287 ± 0%  -1.33%  (p=0.000 n=17+17)
```
- third commit
```
name                                          old time/op    new time/op    delta
FlowSetup/vectorize=true/distribute=true-24     19.0µs ± 5%    19.0µs ± 4%    ~     (p=0.940 n=19+18)
FlowSetup/vectorize=true/distribute=false-24    16.7µs ± 3%    16.7µs ± 3%    ~     (p=0.763 n=20+20)

name                                          old alloc/op   new alloc/op   delta
FlowSetup/vectorize=true/distribute=true-24     33.2kB ± 0%    33.1kB ± 0%  -0.06%  (p=0.000 n=16+16)
FlowSetup/vectorize=true/distribute=false-24    29.7kB ± 1%    29.7kB ± 0%  -0.17%  (p=0.000 n=19+16)

name                                          old allocs/op  new allocs/op  delta
FlowSetup/vectorize=true/distribute=true-24        318 ± 0%       317 ± 0%  -0.31%  (p=0.000 n=20+19)
FlowSetup/vectorize=true/distribute=false-24       287 ± 0%       286 ± 0%  -0.39%  (p=0.000 n=17+17)
```
All commits together:
```
name                                           old time/op    new time/op    delta
FlowSetup/vectorize=true/distribute=true-24      18.9µs ± 3%    18.8µs ± 4%    ~     (p=0.422 n=18+20)
FlowSetup/vectorize=true/distribute=false-24     16.7µs ± 3%    16.7µs ± 2%    ~     (p=0.971 n=19+19)
FlowSetup/vectorize=false/distribute=true-24     18.0µs ± 2%    18.1µs ± 5%    ~     (p=0.664 n=20+20)
FlowSetup/vectorize=false/distribute=false-24    16.1µs ± 3%    16.1µs ± 2%    ~     (p=0.862 n=19+20)

name                                           old alloc/op   new alloc/op   delta
FlowSetup/vectorize=true/distribute=true-24      33.3kB ± 0%    33.2kB ± 0%  -0.49%  (p=0.000 n=19+20)
FlowSetup/vectorize=true/distribute=false-24     29.9kB ± 0%    29.7kB ± 0%  -0.48%  (p=0.000 n=16+19)
FlowSetup/vectorize=false/distribute=true-24     33.8kB ± 0%    33.8kB ± 0%  -0.05%  (p=0.000 n=17+18)
FlowSetup/vectorize=false/distribute=false-24    30.4kB ± 0%    30.3kB ± 0%  -0.06%  (p=0.000 n=18+18)

name                                           old allocs/op  new allocs/op  delta
FlowSetup/vectorize=true/distribute=true-24         324 ± 0%       318 ± 0%  -1.85%  (p=0.000 n=19+19)
FlowSetup/vectorize=true/distribute=false-24        293 ± 0%       287 ± 0%  -2.05%  (p=0.000 n=17+17)
FlowSetup/vectorize=false/distribute=true-24        289 ± 0%       288 ± 0%  -0.35%  (p=0.000 n=20+20)
FlowSetup/vectorize=false/distribute=false-24       258 ± 0%       257 ± 0%  -0.39%  (p=0.000 n=20+20)
```