@yuzefovich 2020-11-10
I personally don't like the third commit - it doesn't give us much improvement but complicates the flow setup story a bit.